+++ src/main/java/me/gfumagali/btgpedidos/repository/ClientOrderRepository.java
@@ -3,5 +3,5 @@ package me.gfumagali.btgpedidos.repository;
 import me.gfumagali.btgpedidos.repository.model.ClientOrders;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-public interface ClientOrderRepository extends MongoRepository<ClientOrders, Integer> {
+public interface ClientOrderRepository extends MongoRepository<ClientOrders, Long> {
 }
